Use references to mutable state in commands (#37)
This pull request includes several changes to the `mneme` Rust project,
focusing on updating method signatures and improving state handling. The
most important changes include modifying the `set_state` method to take
a reference and updating the `apply` method to return a reference to
`self`.

Version update:
*
[`Cargo.toml`](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542L3-R3):
Updated the version of the `mneme` package from `0.4.2` to `0.5.0`.

Method signature updates:
*
[`src/command.rs`](diffhunk://#diff-de6b9cfd1a256da7900d35e26aa7279bd713110e68e91b52616befb7397848dfL17-R17):
Changed the `set_state` method in the `Command` trait to take a
reference to `Self::State`.
*
[`src/lib.rs`](diffhunk://#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759L177-R177):
Updated the `set_state` methods in test implementations to take a
reference to `Self::State`.
[[1]](diffhunk://#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759L177-R177)
[[2]](diffhunk://#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759L337-R338)
[[3]](diffhunk://#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759L479-R478)
*
[`tests/test_cases.rs`](diffhunk://#diff-4300fcbb75cb722103f54594a5422baf31a5e7753a5ad32e7f9eb2ecdf9b8c83L36-R36):
Modified the `set_state` methods in various command implementations to
take a reference to `Self::State`.
[[1]](diffhunk://#diff-4300fcbb75cb722103f54594a5422baf31a5e7753a5ad32e7f9eb2ecdf9b8c83L36-R36)
[[2]](diffhunk://#diff-4300fcbb75cb722103f54594a5422baf31a5e7753a5ad32e7f9eb2ecdf9b8c83L66-R66)
[[3]](diffhunk://#diff-4300fcbb75cb722103f54594a5422baf31a5e7753a5ad32e7f9eb2ecdf9b8c83L95-R95)
[[4]](diffhunk://#diff-4300fcbb75cb722103f54594a5422baf31a5e7753a5ad32e7f9eb2ecdf9b8c83L147-R147)

State handling improvements:
*
[`src/command.rs`](diffhunk://#diff-de6b9cfd1a256da7900d35e26aa7279bd713110e68e91b52616befb7397848dfL39-R45):
Updated the `apply` method in the `AggregateState` trait to return a
reference to `self` and modified the implementation for the unit type.
*
[`src/lib.rs`](diffhunk://#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759L381-R391):
Changed the `apply` method in the `StatefulCommandState` implementation
to return a reference to `self`.
*
[`tests/test_cases.rs`](diffhunk://#diff-4300fcbb75cb722103f54594a5422baf31a5e7753a5ad32e7f9eb2ecdf9b8c83L105-R115):
Updated the `apply` method in the `StatefulCommandState` implementation
to return a reference to `self`.